fix: Don't insert PTO for non-traceable transactions
