+++ apps/explorer/lib/explorer/chain/pending_operations_helper.ex
@@ -115,8 +115,10 @@ defmodule Explorer.Chain.PendingOperationsHelper do
         pto_params =
           Transaction
           |> where([t], t.block_number in ^pbo_block_numbers)
-          |> select([t], %{transaction_hash: t.hash})
+          |> select([t], %{hash: t.hash, type: t.type})
           |> Repo.all()
+          |> Transaction.filter_non_traceable_transactions()
+          |> Enum.map(&%{transaction_hash: &1.hash})
           |> Helper.add_timestamps()
 
         Repo.insert_all(PendingTransactionOperation, pto_params, on_conflict: :nothing)
