+++ apps/explorer/test/explorer/chain/transaction_test.exs
@@ -1028,4 +1028,17 @@ defmodule Explorer.Chain.TransactionTest do
                |> Enum.reverse()
     end
   end
+
+  describe "filter_non_traceable_transactions/1" do
+    test "does not raise when transaction params do not include type on zetachain" do
+      chain_type = Application.get_env(:explorer, :chain_type)
+      Application.put_env(:explorer, :chain_type, :zetachain)
+
+      on_exit(fn -> Application.put_env(:explorer, :chain_type, chain_type) end)
+
+      transaction_params = %{block_number: 13_393_871, hash: "0x123", index: 427}
+
+      assert [transaction_params] == Transaction.filter_non_traceable_transactions([transaction_params])
+    end
+  end
 end
